// src/app/model/action-type.ts
export enum ActionTypes {
    Register = '[Login] Register User',
    RegisterSuccess = '[Login] Register Success',
    RegisterError = '[Login] Register Error',
    Login = '[Login] Login',
    LoginSuccess = '[Login] Login Success',
    LoginError = '[Login] Login Error',
    GetLoggedInUser = '[Login] Get Logged In User',
    GetLoggedInUserSuccess = '[Login] Get Logged In User',
    Logout = '[Login] Logout',
    LogoutSuccess = '[Login] Logout Success'
  }